pyproject.toml
[project]
name = "jpgs-to-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pymupdf>=1.26.6",
]
